1.9??
svn path=/trunk/; revision=195

+++ src/gnet/geventloop_win32.cpp
@@ -1,5 +1,5 @@
 //
-// Copyright (C) 2001-2009 Graeme Walker <graeme_walker@users.sourceforge.net>
+// Copyright (C) 2001-2010 Graeme Walker <graeme_walker@users.sourceforge.net>
 // 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -411,7 +411,7 @@ void GNet::Winsock::setTimeout( G::DateTime::EpochTime t , bool & )
 	if( t != 0U )
 	{
 		G::DateTime::EpochTime now = G::DateTime::now() ;
-		unsigned int interval = t > now ? (t - now) : 0U ;
+		unsigned int interval = t > now ? static_cast<unsigned int>(t-now) : 0U ;
 		unsigned long ms = interval ;
 		ms *= 1000UL ;
 		G_DEBUG( "GNet::Winsock::setTimeout: SetTimer(): " << ms << "ms" ) ;
